feat: draw author signature text (not working!)
I DID NOT SOLVE THE PROBLEM!
Reference: https://onlinepngtools.com/create-transparent-png

I typed the author signature text in Word and took a screenshot of it. Then I used an online PNG tool to make the white background transparent. Instead of proper text, the texture that's being shown are elongated lines.