feat: use blinn-phong lighting model, not phong
Reference: https://learnopengl.com/Advanced-Lighting/Advanced-Lighting

The Blinn-Phong lighting model produces more realistic results than the Phong lighting model. It does so by using the halfway vector between the light's direction and the view direction. This ensures that the angle, which cosine value is used as the base of the specular lighting component intensity exponentiation, never has a value higher than 90 degrees. In other words, the specular area will not be cut off unless the light source is below the surface.