fix: vertices of back side of groceries in ccw
In order for the face culling to work properly, the vertices of every polygon have to be defined in the counter-clockwise winding order. In other words, vertices should be defined as if the camera is looking at them. Previously, I did not define the vertices of the back side of the groceries inside the refrigerator in CCW order. This resulted in back side of groceries not being visible when camera was looking at them.

Reference: https://learnopengl.com/Advanced-OpenGL/Face-culling