fix: rotate door around its right edge
The rotation axis needs to be moved to the right side of the refrigerator door. By the try-out method, I determined that this the rotation axis moves in the direction opposite of what is specified in the translation vector. Therefore, the x-value of translation vector needs to be half of door's width.